refactor: reorganize and clean up codebase
- Bump version in Cargo.toml to 0.4.12-alpha.1
- Remove unused Position enum in config.rs
- Rename types.rs to gmatch.rs and move GMatch-related code
- Refactor GMatch to include to_widget method for converting to gtk::Widget
- Clean up imports and remove unused dependencies
- Refactor main.rs:
  - Simplify module imports and reorganize main function
  - Remove unused fork and copy functions
  - Simplify GTK widget setup and event handling
- Refactor plugins.rs:
  - Remove build_label and build_image functions, moved to utils.rs
  - Simplify handle_matches and refresh_matches functions
- Refactor ui.rs:
  - Modularize window and entry setup functions
  - Simplify event handler connections
  - Consolidate refresh_matches call

Overall improvements for readability, maintainability, and performance.
