+++ anyrun/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "anyrun"
-version = "0.4.12-alpha"
+version = "0.4.12-alpha.1"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
