+++ examples/config.ron
@@ -4,12 +4,17 @@ Config(
     // Fraction(n): A fraction of the width or height of the full screen (depends on exclusive zones and the settings related to them) window respectively
 
     // The width of the runner
-    width: Absolute(800),
+    width: Fraction(0.5),
 
     // The height of the runner
-    height: Absolute(0),
+    height: Fraction(0.5),
 
     // Array of edges where to anchor window. Window will be stretched if two opposite edges specifyed
+    // You can specify empty array to center the window like th follows
+    // ```
+    //     edges: [],
+    // ```
+    // But if you delete/comment this option if will have default value
     // Possible values: Left, Right, Top, Bottom
     // Default: [Top]
     edges: [Top],
